+++ BlueWizard/Buffer.m
@@ -2,7 +2,7 @@
 #import "Autocorrelator.h"
 
 @interface Buffer ()
-@property (nonatomic) float *samples;
+@property (nonatomic) double *samples;
 @property (nonatomic) NSUInteger size;
 @property (nonatomic) NSUInteger sampleRate;
 @end
@@ -14,12 +14,12 @@
     if (self = [super init]) {
         self.size       = size;
         self.sampleRate = sampleRate;
-        self.samples = malloc(sizeof(float) * self.size);
+        self.samples = malloc(sizeof(double) * self.size);
     }
     return self;
 }
 
--(instancetype)initWithSamples:(float *)samples
+-(instancetype)initWithSamples:(double *)samples
                           size:(NSUInteger)size
                     sampleRate:(NSUInteger)sampleRate {
     if (self = [self initWithSize:size sampleRate:sampleRate]) {
@@ -28,7 +28,7 @@
     return self;
 }
 
--(void)copySamples:(float *)samples {
+-(void)copySamples:(double *)samples {
     for (int i = 0; i < self.size; i++) {
         self.samples[i] = samples[i];
     }
@@ -38,7 +38,7 @@
     free(self.samples);
 }
 
--(float)energy {
+-(double)energy {
     return [Autocorrelator sumOfSquaresFor:self];
 }
