+++ BlueWizard/HammingWindow.m
@@ -5,7 +5,7 @@
 
 +(void)processBuffer:(Buffer *)buffer {
     for (int i=0; i < buffer.size; i++) {
-        float window = 0.54f - 0.46f * cos(2 * M_PI * i / (buffer.size - 1));
+        double window = 0.54f - 0.46f * (double long)cos(2 * M_PI * i / (buffer.size - 1));
         buffer.samples[i] *= window;
     }
 }
